Implement power operator
I'm not sure if we should do this, as unlike the other operators that
are simply doing the equivalent operation as they would on an int, in
this case we are repurposing an operator '^' that, in int world, is the
excusive or (XOR) operator. Having it do something completely different
here may confuse users, and it doesn't help that ints and rationals may
end up being used in similar places.