Implement a try pattern conversion to int
Although I'm not a big fan of the try-parse, out parameter pattern and
would prefer to use a method returning an Either<int, Rational> to be
nice and functional, I don't want to have something as low level as a
rational number have a dependency on a functional library like
language-ext (in fact, it really shouldn't depend on anything outside
the core libraries). We can always add them later as extension methods,
separate to the struct itself.