Add new way to create diagonal matrices
Also as part of this ran into issues around Rational numbers:
initialising arrays of them resulted in them using the default struct
initialisation and giving arrays full of 0/0 ratios, which is a bad
thing. This is fundamentally unavoidable as far as I can see, but I've
added some helper methods to hopefully ameliorate the issue.